Allure report corrections
