+++ pom.xml
@@ -3,12 +3,10 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
     <groupId>com.example</groupId>
     <artifactId>Sprint_3</artifactId>
     <version>1.0-SNAPSHOT</version>
     <name>Sprint_3</name>
-
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.target>11</maven.compiler.target>
@@ -17,8 +15,6 @@
         <aspectj.version>1.9.9.1</aspectj.version>
         <allure.version>2.20.0</allure.version>
     </properties>
-
-
     <dependencies>
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
